【CINN】move ir_replace from cinn/optim to cinn/ir/utils (#57524)
* move ir_replace from cinn/optim to cinn/ir/utils

* delete extra modification
