+++ paddle/cinn/optim/eliminate_broadcast_in_forloop.cc
@@ -19,8 +19,8 @@
 
 #include "paddle/cinn/ir/utils/ir_mutator.h"
 #include "paddle/cinn/ir/utils/ir_printer.h"
+#include "paddle/cinn/ir/utils/ir_replace.h"
 #include "paddle/cinn/ir/utils/ir_visitor.h"
-#include "paddle/cinn/optim/ir_replace.h"
 
 namespace cinn {
 namespace optim {
@@ -54,7 +54,7 @@ struct EliminateBroadcastInForloop : public ir::IRMutator<Expr*> {
       std::tie(let_expr, tmp) = CreateTmpLet(broadcast);
       let_exprs.push_back(let_expr);
 
-      optim::IrReplace(expr, broadcast, tmp);
+      cinn::ir::ir_utils::IrReplace(expr, broadcast, tmp);
     }
 
     // insert the let expressions to the outer forloop.
